requirements.txt
# Library dependencies
# pip install -r requirements.txt

#### Required for operation ####
git+https://eljef.dev/python/eljef_core.git@2023.06.1
mutagen
Pillow
rgain3

#### Needed for building ####
build
installer
wheel

#### Needed for linting ####
flake8
pylint

#### Needed for testing ####
pytest

#### Needed for testing coverage ****
pytest-cov
